Reorganize tabular
