+++ fastai/tabular/__init__.py
@@ -3,7 +3,8 @@ from ..basics import *
 from .data import *
 from .transform import *
 from .models import *
+from .learner import *
 from .. import tabular
 
-__all__ = [*basics.__all__, *data.__all__, *transform.__all__, *models.__all__, 'tabular']
+__all__ = [*basics.__all__, *data.__all__, *transform.__all__, *models.__all__, *learner.__all__, 'tabular']
